get transaction works

+++ wallet/src/main/kotlin/org/trustnote/db/entity/Outputs.java
@@ -50,7 +50,7 @@ public class Outputs extends TBaseEntity {
     @ColumnInfo(
             name = "amount"
     )
-    public int amount;
+    public long amount;
 
     @ColumnInfo(
             name = "blinding"
